refactor(map): extract JoinSea passes into standalone MapGenPass classes (#3588)
Split the former multi-pass `_TileMapGenJoinSea` family (a `part of`
`tile_map_generator.dart`, exempt from the uniform `MapGenPass` protocol)
into three standalone single-pass `MapGenPass` services:

- `ContinentJoinPass` (Pass 10, ContinentJoinPassPayload -> ContinentJoinPassResult)
- `TerrainJitterPass` (Pass 10b, TerrainJitterPassPayload -> void)
- `SeaZoneSubdividePass` (Pass 11, SeaZoneSubdividePassPayload -> SeaZoneSubdividePassResult)

To allow the extracted passes to name `TileMapParams` without `part`/`part of`
coupling, `TileMapParams` is moved into its own library `tile_map_params.dart`
and re-exported from `tile_map_generator.dart` (no change to the public barrel
surface). The orchestrator now drives all three passes via `run`; the lakes
pass references `ContinentJoinPass` for `preserveSeaFraction`. Generation
output is byte-identical (determinism tests pass).

Tooling/SPEC:
- `check_map_gen_stage_protocol` now binds the three new pass classes (all
  bound families adopt MapGenPass) and tolerates formatter-wrapped
  `implements` declarations.
- Updated `tile-map-gen-algorithm.md` and `repo-lint.md` for the new names.

Tests: per-pass unit tests (positive + negative) for all three passes.

Refs #3588
