+++ packages/colonizethis_map/lib/src/gen/tile_map_gen_continent_join_pass.dart
@@ -1,27 +1,75 @@
-/// Pass 10–11: join continents, terrain jitter, sea subdivision.
+/// Pass 10: join continents by carving shortest sea-cell land bridges.
 ///
-/// The continent-joining, terrain-jitter, and sea-subdivision passes were
-/// previously split across three sibling part fragments; they are consolidated
-/// here as cohesive extensions on [_TileMapGenJoinSea] (Refs #3574, slice 7).
+/// Extracted from the former `_TileMapGenJoinSea` bridge fragment into a
+/// standalone [MapGenPass] family (Refs #3588). Constructor-injected
+/// dependencies replace the previous `part of` library-scope coupling so the
+/// pass is testable in isolation. SPEC/program/tile-map-gen-algorithm.md.
+library;
 
-part of 'tile_map_generator.dart';
+import 'dart:math';
 
-/// Exempt from the uniform [MapGenPass] entry point (Refs #3574, slice 4):
-/// this family owns three heterogeneous passes — continent joining
-/// (grid+terrain+resource in/out), terrain jitter (in-place mutation), and
-/// sea-zone subdivision (grid in/out) — that share no single representative
-/// payload/result shape. It therefore implements [MapGenStage] only; the
-/// orchestrator drives its three passes via their dedicated methods.
-class _TileMapGenJoinSea implements MapGenStage {
-  _TileMapGenJoinSea(this.params, this._log, this._graph);
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_map/package_logger.dart';
+
+import '../tile_map_directions.dart';
+import '../tile_map_grid.dart';
+import 'map_gen_pass_payloads.dart';
+import 'map_gen_stage.dart';
+import 'tile_map_grid_graph.dart';
+import 'tile_map_params.dart';
+import 'tile_map_resource_cap_state.dart';
+import 'tile_map_resource_placement.dart';
+
+/// Pass 10 continent-joining service.
+///
+/// For each continent with more than one land component, connects two
+/// components with a shortest path of sea cells and preserves the overall sea
+/// fraction by restoring an equal number of coastal land tiles to sea.
+class ContinentJoinPass
+    implements MapGenPass<ContinentJoinPassPayload, ContinentJoinPassResult> {
+  ContinentJoinPass(this.params, this._log, this._graph);
 
   @override
   final TileMapParams params;
   final CtLogger _log;
   final TileMapGridGraph _graph;
-}
 
-extension _TileMapGenJoinSeaBridgePart on _TileMapGenJoinSea {
+  /// Uniform pass entry: runs continent joining when [TileMapParams.joinContinents]
+  /// is set, otherwise returns the inputs unchanged (Refs #3588).
+  @override
+  ContinentJoinPassResult run(MapGenPassContext<ContinentJoinPassPayload> ctx) {
+    final payload = ctx.payload;
+    if (!params.joinContinents) {
+      return ContinentJoinPassResult(
+        grid: payload.grid,
+        terrainGrid: payload.terrainGrid,
+        resourceGrid: payload.resourceGrid,
+        didJoin: false,
+      );
+    }
+    final (g, tg, rg, didJoin) = joinContinents(
+      payload.grid,
+      payload.terrainGrid,
+      payload.resourceGrid,
+      payload.provinceToContinent,
+      payload.seaZoneId,
+      payload.mapRegionId,
+      payload.landSeeds,
+      payload.continentBySeedIndex,
+      payload.resourceRules,
+      payload.rnd,
+    );
+    if (didJoin) {
+      ctx.log('Pass 10: Join continents (land bridges added)');
+    }
+    return ContinentJoinPassResult(
+      grid: g,
+      terrainGrid: tg,
+      resourceGrid: rg,
+      didJoin: didJoin,
+    );
+  }
+
   /// Join step: for each continent with >1 land component, connect two by a shortest path of sea cells. Returns (grid, terrainGrid, resourceGrid, didJoin).
   (List<List<String>>, List<List<TerrainType?>>?, List<List<Resource?>>?, bool)
   joinContinents(
@@ -108,11 +156,11 @@ extension _TileMapGenJoinSeaBridgePart on _TileMapGenJoinSea {
           path.length,
           landCellsExcludedFromSeaRestore: bridgeCells,
         );
-        for (final (x, y) in restoredToSea) {
-          for (var ci = 0; ci < numContinents; ci++) {
-            landCellsByContinent[ci].remove((x, y));
-          }
-        }
+        _removeCellsFromAllContinents(
+          landCellsByContinent,
+          restoredToSea,
+          numContinents,
+        );
       }
       if (joinIterations >= maxJoinIterationsPerContinent) {
         final stillSplit = _graph.connectedComponentsOfLand(
@@ -129,8 +177,20 @@ extension _TileMapGenJoinSeaBridgePart on _TileMapGenJoinSea {
     return (g, tg, rg, didJoin);
   }
 
-  int countSeaCells(List<List<String>> grid, String seaZoneId) =>
-      _graph.countSeaCells(grid, seaZoneId);
+  /// Drops every cell in [cells] from each continent's land-cell set (used after
+  /// [preserveSeaFraction] restores coastal land back to sea). Extracted to keep
+  /// [joinContinents] control-flow nesting within policy.
+  void _removeCellsFromAllContinents(
+    List<Set<(int x, int y)>> landCellsByContinent,
+    List<(int x, int y)> cells,
+    int numContinents,
+  ) {
+    for (final (x, y) in cells) {
+      for (var ci = 0; ci < numContinents; ci++) {
+        landCellsByContinent[ci].remove((x, y));
+      }
+    }
+  }
 
   void _applyBridgePathCells(
     List<List<String>> g,
@@ -306,6 +366,9 @@ extension _TileMapGenJoinSeaBridgePart on _TileMapGenJoinSea {
     );
   }
 
+  /// Restore [count] coastal land tiles back to sea (highest ocean adjacency
+  /// first) to preserve the overall sea fraction after a join. Shared with the
+  /// lakes/moats pass. Returns the restored cell coordinates.
   List<(int x, int y)> preserveSeaFraction(
     List<List<String>> grid,
     List<List<TerrainType?>>? terrainGrid,
@@ -337,228 +400,5 @@ extension _TileMapGenJoinSeaBridgePart on _TileMapGenJoinSea {
       restoredToSea.add((x, y));
     }
     return restoredToSea;
-  }
-}
-
-extension _TileMapGenJoinSeaJitterPart on _TileMapGenJoinSea {
-  bool _jitterTileIsTerrainOrProvinceEdge(
-    int x,
-    int y,
-    String provinceId,
-    TerrainType dominant,
-    List<List<String>> grid,
-    List<List<TerrainType?>> terrainGrid,
-    int width,
-    int height,
-    List<(int dx, int dy)> directions4,
-  ) {
-    for (final (dx, dy) in directions4) {
-      final nx = x + dx;
-      final ny = y + dy;
-      if (nx < 0 || nx >= width || ny < 0 || ny >= height) continue;
-      final neighborProvince = grid[ny][nx];
-      final neighborTerrain = terrainGrid[ny][nx];
-      if (neighborProvince != provinceId) return true;
-      if (neighborTerrain != null && neighborTerrain != dominant) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  void jitterTerrainByProvince(
-    List<List<String>> grid,
-    List<List<TerrainType?>> terrainGrid,
-    List<List<Resource?>> resourceGrid,
-    String regionId,
-    Random rnd,
-  ) {
-    final allowedNonMountain = allowedTerrainsForRegion(
-      regionId,
-    ).where((t) => t != TerrainType.mountain).toList();
-    if (allowedNonMountain.isEmpty) return;
-
-    final height = grid.length;
-    if (height == 0) return;
-    final width = grid[0].length;
-    final tilesByProvince = <String, List<(int x, int y)>>{};
-    final provinceIdPattern = RegExp(r'^p\d+$');
-    TileMapGrid.forEachCell(grid, (y, x, id) {
-      if (!provinceIdPattern.hasMatch(id)) return;
-      tilesByProvince.putIfAbsent(id, () => []).add((x, y));
-    });
-    if (tilesByProvince.isEmpty) return;
-
-    const directions4 = kTileMapDirections4;
-    const directions8 = kTileMapDirections8;
-
-    for (final entry in tilesByProvince.entries) {
-      final tiles = entry.value;
-      if (tiles.length < params.jitterMinProvinceSize) continue;
-
-      final counts = <TerrainType, int>{};
-      var terrainTiles = 0;
-      for (final (x, y) in tiles) {
-        final t = terrainGrid[y][x];
-        if (t == null) continue;
-        counts[t] = (counts[t] ?? 0) + 1;
-        terrainTiles++;
-      }
-      if (terrainTiles == 0 || counts.isEmpty) continue;
-
-      TerrainType dominant = counts.keys.first;
-      var maxCount = counts[dominant]!;
-      for (final e in counts.entries) {
-        if (e.value > maxCount) {
-          dominant = e.key;
-          maxCount = e.value;
-        }
-      }
-      final fDom = maxCount / terrainTiles;
-      if (fDom < params.jitterHomogeneityThreshold) continue;
-
-      final candidates = <(int x, int y)>[];
-      for (final (x, y) in tiles) {
-        if (terrainGrid[y][x] != dominant) continue;
-        if (resourceGrid[y][x] != null) continue;
-        if (!_jitterTileIsTerrainOrProvinceEdge(
-          x,
-          y,
-          entry.key,
-          dominant,
-          grid,
-          terrainGrid,
-          width,
-          height,
-          directions4,
-        )) {
-          continue;
-        }
-        candidates.add((x, y));
-      }
-      if (candidates.isEmpty) continue;
-
-      candidates.shuffle(rnd);
-      final maxChanges = (params.jitterMaxFraction * tiles.length).floor();
-      if (maxChanges <= 0) continue;
-      var changes = 0;
-      for (final (x, y) in candidates) {
-        if (changes >= maxChanges) break;
-        if (rnd.nextDouble() > params.jitterProbability) continue;
-
-        final neighborCounts = <TerrainType, int>{};
-        for (final (dx, dy) in directions8) {
-          final nx = x + dx;
-          final ny = y + dy;
-          if (nx < 0 || nx >= width || ny < 0 || ny >= height) continue;
-          if (grid[ny][nx] != entry.key) continue;
-          final nt = terrainGrid[ny][nx];
-          if (nt == null || nt == dominant || nt == TerrainType.mountain) {
-            continue;
-          }
-          neighborCounts[nt] = (neighborCounts[nt] ?? 0) + 1;
-        }
-
-        final supported = neighborCounts.entries
-            .where((e) => e.value >= params.jitterNeighborSupportThreshold)
-            .map((e) => e.key)
-            .toList();
-        if (supported.isEmpty) continue;
-        terrainGrid[y][x] = supported[rnd.nextInt(supported.length)];
-        changes++;
-      }
-    }
-  }
-}
-
-extension _TileMapGenJoinSeaSubdividePart on _TileMapGenJoinSea {
-  /// Pass 11: Subdivide sea so each zone has at most maxSeaZoneFraction * totalSea tiles.
-  /// Returns (newGrid, total sea zone count).
-  (List<List<String>>, int) subdivideSeaZonesWithCap(
-    List<List<String>> grid,
-    String seaZoneId,
-    int totalSea,
-  ) {
-    final components = _graph.connectedComponentsOfSea(grid, seaZoneId);
-    if (components.isEmpty) return (grid, 0);
-    final sorted = List<Set<(int x, int y)>>.from(components)
-      ..sort((a, b) {
-        final (minYa, minXa) = _graph.minYx(a);
-        final (minYb, minXb) = _graph.minYx(b);
-        if (minYa != minYb) return minYa.compareTo(minYb);
-        return minXa.compareTo(minXb);
-      });
-    final g = snapshotGrid(grid);
-    final maxPerZone = (params.maxSeaZoneFraction * totalSea).floor();
-    var nextSeaZoneIndex = 1;
-    for (final component in sorted) {
-      final size = component.length;
-      if (maxPerZone <= 0 || size <= maxPerZone) {
-        final id = 's$nextSeaZoneIndex';
-        nextSeaZoneIndex++;
-        for (final (x, y) in component) {
-          g[y][x] = id;
-        }
-        continue;
-      }
-      final K = (size / maxPerZone).ceil().clamp(1, size);
-      final seeds = _placeSeaSeedsFarthestPoint(component, K);
-      final seedMap = <String, (int x, int y)>{
-        for (var i = 0; i < seeds.length; i++)
-          's${nextSeaZoneIndex + i}': seeds[i],
-      };
-      final assignment = assignCellsToNearestSeed(
-        component,
-        seedMap,
-        noiseScale: params.voronoiNoiseScale,
-        noiseSeed: params.seed,
-      );
-      for (final entry in assignment.entries) {
-        final (x, y) = entry.key;
-        g[y][x] = entry.value;
-      }
-      nextSeaZoneIndex += K;
-    }
-    return (g, nextSeaZoneIndex - 1);
-  }
-
-  /// Place K well-spread seeds in [cells] using farthest-point sampling.
-  List<(int x, int y)> _placeSeaSeedsFarthestPoint(
-    Set<(int x, int y)> cells,
-    int K,
-  ) {
-    if (cells.isEmpty || K <= 0) return [];
-    final list = cells.toList();
-    if (K >= list.length) return list;
-    list.sort((a, b) {
-      if (a.$2 != b.$2) return a.$2.compareTo(b.$2);
-      return a.$1.compareTo(b.$1);
-    });
-    final chosen = <(int x, int y)>[list.first];
-    for (var i = 1; i < K; i++) {
-      chosen.add(_bestFarthestPointCellFromList(list, chosen));
-    }
-    return chosen;
-  }
-
-  (int x, int y) _bestFarthestPointCellFromList(
-    List<(int x, int y)> list,
-    List<(int x, int y)> chosen,
-  ) {
-    var bestCell = list.first;
-    var bestMinD2 = 0;
-    for (final (x, y) in list) {
-      if (chosen.contains((x, y))) continue;
-      var minD2 = kUnsetSquaredDistanceInt31;
-      for (final (sx, sy) in chosen) {
-        final d2 = (x - sx) * (x - sx) + (y - sy) * (y - sy);
-        if (d2 < minD2) minD2 = d2;
-      }
-      if (minD2 > bestMinD2) {
-        bestMinD2 = minD2;
-        bestCell = (x, y);
-      }
-    }
-    return bestCell;
   }
 }
