+++ test/check_map_gen_stage_protocol_test.dart
@@ -8,14 +8,14 @@ void main() {
   group('findMapGenStageProtocolViolations', () {
     test('accepts a service class implementing MapGenStage', () {
       const src = r'''
-class _TileMapGenJoinSea implements MapGenStage {
+class _TileMapGenLakesProvinces implements MapGenStage {
   @override
   final TileMapParams params;
 }
 ''';
       final violations = findMapGenStageProtocolViolations(
         relativePath:
-            'packages/colonizethis_map/lib/src/gen/tile_map_generator_join_sea.dart',
+            'packages/colonizethis_map/lib/src/gen/tile_map_generator_lakes_provinces.dart',
         source: src,
       );
       expect(violations, isEmpty);
@@ -23,13 +23,13 @@ class _TileMapGenJoinSea implements MapGenStage {
 
     test('flags a service class missing MapGenStage implementation', () {
       const src = r'''
-class _TileMapGenJoinSea {
+class _TileMapGenLakesProvinces {
   final TileMapParams params;
 }
 ''';
       final violations = findMapGenStageProtocolViolations(
         relativePath:
-            'packages/colonizethis_map/lib/src/gen/tile_map_generator_join_sea.dart',
+            'packages/colonizethis_map/lib/src/gen/tile_map_generator_lakes_provinces.dart',
         source: src,
       );
       expect(violations, hasLength(1));
@@ -81,15 +81,17 @@ class TileMapGenLandSeeds
   });
 
   group('runCheckMapGenStageProtocol', () {
-    test('passes on the live repository tree (>=3 families adopt MapGenPass)',
-        () {
-      final lines = <String>[];
-      final code = runCheckMapGenStageProtocol(
-        Directory.current.path,
-        info: lines.add,
-        err: lines.add,
-      );
-      expect(code, 0, reason: lines.join('\n'));
-    });
+    test(
+      'passes on the live repository tree (>=3 families adopt MapGenPass)',
+      () {
+        final lines = <String>[];
+        final code = runCheckMapGenStageProtocol(
+          Directory.current.path,
+          info: lines.add,
+          err: lines.add,
+        );
+        expect(code, 0, reason: lines.join('\n'));
+      },
+    );
   });
 }
